+++ lib/rubocop/cop/style/nested_ternary_operator.rb
@@ -6,7 +6,7 @@ module Rubocop
       # This cop checks for nested ternary op expressions.
       class NestedTernaryOperator < Cop
         MSG = 'Ternary operators must not be nested. Prefer if/else ' \
-            'constructs instead.'
+              'constructs instead.'
 
         def on_if(node)
           loc = node.loc