+++ lib/rubocop/cop/style/numeric_literals.rb
@@ -12,7 +12,7 @@ module Rubocop
         include ConfigurableMax
 
         MSG = 'Separate every 3 digits in the integer portion of a number ' \
-          'with underscores(_).'
+              'with underscores(_).'
 
         def on_int(node)
           check(node)