+++ lib/rubocop/cop/style/percent_literal_delimiters.rb
@@ -31,7 +31,7 @@ module Rubocop
           delimiters = preferred_delimiters(type)
 
           "`#{type}`-literals should be delimited by " \
-            "`#{delimiters[0]}` and `#{delimiters[1]}`"
+          "`#{delimiters[0]}` and `#{delimiters[1]}`"
         end
 
         private